Convert code-mode transform to the CodeMode plugin
Moves code_mode out of fastmcp.experimental.transforms into
fastmcp.server.plugins.code_mode and wraps the existing transform in
a CodeMode(Plugin[CodeModeConfig]) class. Package is split into
plugin/transform/discovery/sandbox modules for clarity.

Breaking change at the old import path: `CodeMode` is now the plugin,
not the transform. Callers that composed the transform directly should
import `CodeModeTransform` from the new location. Both continue to
work via the deprecation shim, which emits a FastMCPDeprecationWarning.

Closes FMCP-27.
